functions/update/requirements.txt
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
